% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptAPI.R
\name{rbu_api_detail_hana}
\alias{rbu_api_detail_hana}
\title{向BW报表写入明细数据至hana}
\usage{
rbu_api_detail_hana(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  page_count = 10000
)
}
\arguments{
\item{conn}{连接信息}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{page_count}{每页数量}
}
\value{
返回结果
}
\description{
向BW报表写入明细数据至hana
}
\examples{
rbu_api_detail_hana()
}
