% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptAPI.R
\name{rbu_api_detail_rds}
\alias{rbu_api_detail_rds}
\title{将过程表写入至RDS}
\usage{
rbu_api_detail_rds(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 6)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
将过程表写入至RDS
}
\examples{
rbu_api_detail_rds()
}
