% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptAPI.R
\name{rbu_api_res_hana}
\alias{rbu_api_res_hana}
\title{向BW报表写入结果数据至hana}
\usage{
rbu_api_res_hana(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 6,
  view_name = "mrpt3_vw_FI_RPA"
)
}
\arguments{
\item{conn}{连接信息}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{view_name}{mrpt3_vw_FI_RPA mrpt2_vw_FI_RPA}
}
\value{
返回结果
}
\description{
向BW报表写入结果数据至hana
}
\examples{
rbu_api_res_hana()
}
