% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptAPI.R
\name{rbu_api_res_rds}
\alias{rbu_api_res_rds}
\title{写入结果表至RDS
T_FI_RPA
rds_vw_T_FI_RPA_all}
\usage{
rbu_api_res_rds(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 6)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
写入结果表至RDS
T_FI_RPA
rds_vw_T_FI_RPA_all
}
\examples{
rbu_api_res_rds()
}
