% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCheck.R
\name{rbu_autoCheck_rpa_vs_target}
\alias{rbu_autoCheck_rpa_vs_target}
\title{管报自动化处理差异表}
\usage{
rbu_autoCheck_rpa_vs_target(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 7,
  FBrand = "自然堂",
  FChannel = "美妆",
  rpa_table = "T_FI_RPA"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{rpa_table}{rpa表名}
}
\value{
返回值
}
\description{
管报自动化处理差异表
}
\examples{
library(rbupkg)
rbu_autoCheck_rpa_vs_target() -> mydata
View(mydata)
}
