% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_ALL}
\alias{rbu_calc_level_ALL}
\title{计算结果表强制为万元
vw_mrpt_res_level_ALL}
\usage{
rbu_calc_level_ALL(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算结果表强制为万元
vw_mrpt_res_level_ALL
}
\examples{
rbu_calc_level_ALL()
}
