% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_ALL3}
\alias{rbu_calc_level_ALL3}
\title{计算结果增加了执行预算数据
vw_mrpt_res_level_ALL3}
\usage{
rbu_calc_level_ALL3(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算结果增加了执行预算数据
vw_mrpt_res_level_ALL3
}
