% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_ALL4}
\alias{rbu_calc_level_ALL4}
\title{计算结果增加了同比数据,19与20年数据
vw_mrpt_res_level_ALL4}
\usage{
rbu_calc_level_ALL4()
}
\value{

}
\description{
计算结果增加了同比数据,19与20年数据
vw_mrpt_res_level_ALL4
}
\examples{
rbu_calc_level_ALL4()
}
