% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_ALL_original}
\alias{rbu_calc_level_ALL_original}
\title{结果表的处理内容，扣除报表项目为空的问题
vw_mrpt_res_level_ALL_original}
\usage{
rbu_calc_level_ALL_original(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
结果表的处理内容，扣除报表项目为空的问题
vw_mrpt_res_level_ALL_original
}
