% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_ALL_zero}
\alias{rbu_calc_level_ALL_zero}
\title{按报表项目进行张量补全}
\usage{
rbu_calc_level_ALL_zero(variables)
}
\arguments{
\item{variables}{vw_mrpt_res_level_ALL_zero}
}
\value{

}
\description{
按报表项目进行张量补全
}
\examples{
rbu_calc_level_ALL_zero()
}
