% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_I07_profitRate}
\alias{rbu_calc_level_I07_profitRate}
\title{计算指标I07}
\usage{
rbu_calc_level_I07_profitRate(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算指标I07
}
\examples{
rbu_calc_level_I07_profitRate()
}
