% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_I08_salesFee}
\alias{rbu_calc_level_I08_salesFee}
\title{计算指标I08}
\usage{
rbu_calc_level_I08_salesFee(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算指标I08
}
\examples{
rbu_calc_level_I08_salesFee()
}
