% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_I26_Salesprofit}
\alias{rbu_calc_level_I26_Salesprofit}
\title{计算指标I26}
\usage{
rbu_calc_level_I26_Salesprofit(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算指标I26
}
\examples{
rbu_calc_level_I26_Salesprofit()
}
