% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_I27_marketFee}
\alias{rbu_calc_level_I27_marketFee}
\title{计算指标I27}
\usage{
rbu_calc_level_I27_marketFee(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算指标I27
}
\examples{
rbu_calc_level_I27_marketFee()
}
