% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_I44_marketProfit}
\alias{rbu_calc_level_I44_marketProfit}
\title{计算指标I44}
\usage{
rbu_calc_level_I44_marketProfit(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
计算指标I44
}
\examples{
rbu_calc_level_I44_marketProfit()
}
