% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_level1}
\alias{rbu_calc_level_level1}
\title{计算结果从明细表到汇总表部分}
\usage{
rbu_calc_level_level1(variables)
}
\arguments{
\item{variables}{变量}
}
\value{
返回值
}
\description{
计算结果从明细表到汇总表部分
}
\examples{
rbu_calc_level_level1()
}
