% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptCalculation.R
\name{rbu_calc_level_level12}
\alias{rbu_calc_level_level12}
\title{结果表，增加了零售额及回款信息
使用得不是很规范，}
\usage{
rbu_calc_level_level12(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
结果表，增加了零售额及回款信息
使用得不是很规范，
}
