% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllcation}
\alias{rbu_feeAllcation}
\title{1、费用分配明细表
rbu_feeAllcation_all_Allocated 同名的函数}
\usage{
rbu_feeAllcation(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
1、费用分配明细表
rbu_feeAllcation_all_Allocated 同名的函数
}
\examples{
rbu_feeAllcation()
}
