% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllcation_all_Allocated}
\alias{rbu_feeAllcation_all_Allocated}
\title{费用分析明细表
mrpt2_vw_ds_all_Allocated 视图
mrpt2_t_ds_all_Allocated  表}
\usage{
rbu_feeAllcation_all_Allocated(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
费用分析明细表
mrpt2_vw_ds_all_Allocated 视图
mrpt2_t_ds_all_Allocated  表
}
\examples{
rbu_feeAllcation_all_Allocated()
}
