% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllcation_detail}
\alias{rbu_feeAllcation_detail}
\title{费用分配明细表
mrpt2_vw_ds_all_Allocated}
\usage{
rbu_feeAllcation_detail(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用分配明细表
mrpt2_vw_ds_all_Allocated
}
\examples{
rbu_feeAllcation_detail()
}
