% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeAllocation_all_TeTongTotal_detail}
\alias{rbu_feeAllocation_all_TeTongTotal_detail}
\title{费用分配处理特通数据
mrpt2_vw_ds_all_TeTongTotal_detail}
\usage{
rbu_feeAllocation_all_TeTongTotal_detail(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用分配处理特通数据
mrpt2_vw_ds_all_TeTongTotal_detail
}
\examples{
rbu_feeAllocation_TeTong()
}
