% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_Biorrier_minusRevenue}
\alias{rbu_feeAllocation_Biorrier_minusRevenue}
\title{费用分配药房冲减收入
mrpt2_vw_ds_all_biorrierYaoFang_RevenueMinus}
\usage{
rbu_feeAllocation_Biorrier_minusRevenue()
}
\value{
返回值
}
\description{
费用分配药房冲减收入
mrpt2_vw_ds_all_biorrierYaoFang_RevenueMinus
}
\examples{
rbu_feeAllocation_Biorrier_minusRevenue()
}
