% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_cost}
\alias{rbu_feeAllocation_cost}
\title{费用分配市场费用}
\usage{
rbu_feeAllocation_cost(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用分配市场费用
}
\examples{
rbu_feeAllocation_cost()
}
