% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_costTotal}
\alias{rbu_feeAllocation_costTotal}
\title{品牌所有渠道营业成本汇总表
vw_mrpt_res_costTotal}
\usage{
rbu_feeAllocation_costTotal(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
品牌所有渠道营业成本汇总表
vw_mrpt_res_costTotal
}
\examples{
rbu_feeAllocation_costTotal()
}
