% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_ka}
\alias{rbu_feeAllocation_ka}
\title{费用分配明细KA数据
mrpt2_vw_ds_all_kaTotal_detail}
\usage{
rbu_feeAllocation_ka(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用分配明细KA数据
mrpt2_vw_ds_all_kaTotal_detail
}
\examples{
rbu_feeAllocation_ka()
}
