% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_market_toChannel}
\alias{rbu_feeAllocation_market_toChannel}
\title{市场费用已分配到各渠道
vw_mrpt_res_MartetFee_toChannel}
\usage{
rbu_feeAllocation_market_toChannel()
}
\value{
返回值
}
\description{
市场费用已分配到各渠道
vw_mrpt_res_MartetFee_toChannel
}
\examples{

rbu_feeAllocation_market_toChannel()
}
