% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_marketRatio}
\alias{rbu_feeAllocation_marketRatio}
\title{市场费用分配率
vw_mrpt_res_marketRatio}
\usage{
rbu_feeAllocation_marketRatio(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
市场费用分配率
vw_mrpt_res_marketRatio
}
\examples{
rbu_feeAllocation_marketRatio()
}
