% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeAllocation.R
\name{rbu_feeAllocation_marktNot}
\alias{rbu_feeAllocation_marktNot}
\title{非市场费用
mrpt2_vw_ds_all_unAllocated_marketNot}
\usage{
rbu_feeAllocation_marktNot()
}
\value{
返回值
}
\description{
非市场费用
mrpt2_vw_ds_all_unAllocated_marketNot
}
\examples{
rbu_feeAllocation_marktNot()
}
