% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection}
\alias{rbu_feeCollection}
\title{费用归集所有明细
mrpt2_vw_ds_all_unAllocated}
\usage{
rbu_feeCollection(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用归集所有明细
mrpt2_vw_ds_all_unAllocated
}
\examples{
rbu_feeCollection_detail()
}
