% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_adj}
\alias{rbu_feeCollection_adj}
\title{费用归集-手调凭证
mrpt2_vw_ds_adj_all}
\usage{
rbu_feeCollection_adj(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用归集-手调凭证
mrpt2_vw_ds_adj_all
}
\examples{
rbu_feeCollection_adj()
}
