% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_adj_all}
\alias{rbu_feeCollection_adj_all}
\title{Title}
\usage{
rbu_feeCollection_adj_all(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
\examples{
rbu_feeCollection_adj_all()
}
