% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_all_biorrierYaoFang_RevenueMinus}
\alias{rbu_feeCollection_all_biorrierYaoFang_RevenueMinus}
\title{费用分配药房冲减收入
mrpt2_vw_ds_all_biorrierYaoFang_RevenueMinus}
\usage{
rbu_feeCollection_all_biorrierYaoFang_RevenueMinus()
}
\value{
返回值
}
\description{
费用分配药房冲减收入
mrpt2_vw_ds_all_biorrierYaoFang_RevenueMinus
}
\examples{
rbu_feeCollection_all_biorrierYaoFang_RevenueMinus()
}
