% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_all_unAllocated}
\alias{rbu_feeCollection_all_unAllocated}
\title{费用归集明细表
mrpt2_vw_ds_all_unAllocated}
\usage{
rbu_feeCollection_all_unAllocated(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
费用归集明细表
mrpt2_vw_ds_all_unAllocated
}
\examples{
rbu_feeCollection_all_unAllocated()
}
