% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_all_unAllocated_market}
\alias{rbu_feeCollection_all_unAllocated_market}
\title{市场费用待分配
mrpt2_vw_ds_all_unAllocated_market}
\usage{
rbu_feeCollection_all_unAllocated_market(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
市场费用待分配
mrpt2_vw_ds_all_unAllocated_market
}
