% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_all_unAllocated_withSuperCustomers}
\alias{rbu_feeCollection_all_unAllocated_withSuperCustomers}
\title{费用待分配包括KA，特通及大客户----
本质在处理第一级与第二级的费用问题
mrpt2_vw_ds_all_unAllocated_withSuperCustomers}
\usage{
rbu_feeCollection_all_unAllocated_withSuperCustomers(variables)
}
\arguments{
\item{variables}{}
}
\value{
返回值
}
\description{
费用待分配包括KA，特通及大客户----
本质在处理第一级与第二级的费用问题
mrpt2_vw_ds_all_unAllocated_withSuperCustomers
}
\examples{
rbu_feeCollection_all_unAllocated_withSuperCustomers()
}
