% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_all_unAllocated_withSuperCustomers_brandChannel}
\alias{rbu_feeCollection_all_unAllocated_withSuperCustomers_brandChannel}
\title{匹配了品牌渠道费用及品牌渠道基础表}
\usage{
rbu_feeCollection_all_unAllocated_withSuperCustomers_brandChannel(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
匹配了品牌渠道费用及品牌渠道基础表
}
\examples{
rbu_feeCollection_all_unAllocated_withSuperCustomers_brandChannel()
}
