% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_bw}
\alias{rbu_feeCollection_bw}
\title{费用归集-BW
mrpt2_vw_ds_bw_all}
\usage{
rbu_feeCollection_bw()
}
\value{
返回值
}
\description{
费用归集-BW
mrpt2_vw_ds_bw_all
}
\examples{
rbu_feeCollection_bw()
}
