% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_bw_all}
\alias{rbu_feeCollection_bw_all}
\title{Title
mrpt2_vw_ds_bw_all}
\usage{
rbu_feeCollection_bw_all(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
mrpt2_vw_ds_bw_all
}
