% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_bw_rpa_ruled}
\alias{rbu_feeCollection_bw_rpa_ruled}
\title{BW写入的结果表
这应该是BW核对核法写入的结果
rds_t_mrpt_ds_bw_rpa_ruled}
\usage{
rbu_feeCollection_bw_rpa_ruled(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
BW写入的结果表
这应该是BW核对核法写入的结果
rds_t_mrpt_ds_bw_rpa_ruled
}
\examples{
rbu_feeCollection_bw_rpa_ruled()
}
