% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap}
\alias{rbu_feeCollection_sap}
\title{费用归集-SAP
mrpt2_vw_ds_sap_all}
\usage{
rbu_feeCollection_sap()
}
\value{
返回值
}
\description{
费用归集-SAP
mrpt2_vw_ds_sap_all
}
\examples{
rbu_feeCollection_sap()
}
