% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_all}
\alias{rbu_feeCollection_sap_all}
\title{1.1.01费用归集SAP凭证
mrpt2_vw_ds_sap_all}
\usage{
rbu_feeCollection_sap_all(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
1.1.01费用归集SAP凭证
mrpt2_vw_ds_sap_all
}
\examples{
rbu_feeCollection_sap_all()
}
