% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_BrandChannel_chandoEcom}
\alias{rbu_feeCollection_sap_BrandChannel_chandoEcom}
\title{Title
mrpt2_vw_data_sap_BrandChannel_chandoEcom}
\usage{
rbu_feeCollection_sap_BrandChannel_chandoEcom(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
mrpt2_vw_data_sap_BrandChannel_chandoEcom
}
\examples{
rbu_feeCollection_sap_BrandChannel_chandoEcom()
}
