% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_BrandChannel_GeneralOthers}
\alias{rbu_feeCollection_sap_BrandChannel_GeneralOthers}
\title{除自然堂电商的费用
mrpt2_vw_data_sap_BrandChannel_GeneralOthers}
\usage{
rbu_feeCollection_sap_BrandChannel_GeneralOthers(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
除自然堂电商的费用
mrpt2_vw_data_sap_BrandChannel_GeneralOthers
}
