% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_BrandChannel_notnull}
\alias{rbu_feeCollection_sap_BrandChannel_notnull}
\title{Title}
\usage{
rbu_feeCollection_sap_BrandChannel_notnull(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
\examples{
rbu_feeCollection_sap_BrandChannel_notnull()
}
