% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_FeeName_generalOthers_notNull}
\alias{rbu_feeCollection_sap_FeeName_generalOthers_notNull}
\title{Title
mrpt2_vw_data_sap_FeeName_generalOthers_notNull}
\usage{
rbu_feeCollection_sap_FeeName_generalOthers_notNull(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
mrpt2_vw_data_sap_FeeName_generalOthers_notNull
}
\examples{
rbu_feeCollection_sap_FeeName_generalOthers_notNull()
}
