% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_rptItem_all}
\alias{rbu_feeCollection_sap_rptItem_all}
\title{1.1.01.03、SAP凭证项目-----------}
\usage{
rbu_feeCollection_sap_rptItem_all()
}
\value{

}
\description{
1.1.01.03、SAP凭证项目-----------
}
\examples{
rbu_feeCollection_sap_rptItem_all()
}
