% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_rptItem_all_allocated}
\alias{rbu_feeCollection_sap_rptItem_all_allocated}
\title{SAP费用分配表
mrpt2_vw_data_sap_rptItem_all_allocated}
\usage{
rbu_feeCollection_sap_rptItem_all_allocated(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
SAP费用分配表
mrpt2_vw_data_sap_rptItem_all_allocated
}
\examples{
rbu_feeCollection_sap_rptItem_all_allocated()
}
