% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_rptItem_all_shared}
\alias{rbu_feeCollection_sap_rptItem_all_shared}
\title{SAP凭证共享成本中心费用待分配
mrpt2_vw_data_sap_rptItem_all_shared}
\usage{
rbu_feeCollection_sap_rptItem_all_shared()
}
\value{

}
\description{
SAP凭证共享成本中心费用待分配
mrpt2_vw_data_sap_rptItem_all_shared
}
\examples{
rbu_feeCollection_sap_rptItem_all_shared()
}
