% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_rptItem_GeneralOthers_notNull}
\alias{rbu_feeCollection_sap_rptItem_GeneralOthers_notNull}
\title{除自然堂电商外的内容
mrpt2_vw_data_sap_rptItem_GeneralOthers_notNull}
\usage{
rbu_feeCollection_sap_rptItem_GeneralOthers_notNull(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
除自然堂电商外的内容
mrpt2_vw_data_sap_rptItem_GeneralOthers_notNull
}
