% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeCollection.R
\name{rbu_feeCollection_sap_rptItem_shared_allocated}
\alias{rbu_feeCollection_sap_rptItem_shared_allocated}
\title{SAP 凭证费用-分配部分
mrpt2_vw_data_sap_salesFee_allocated}
\usage{
rbu_feeCollection_sap_rptItem_shared_allocated(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
SAP 凭证费用-分配部分
mrpt2_vw_data_sap_salesFee_allocated
}
\examples{
rbu_feeCollection_sap_rptItem_shared_allocated()
}
