% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_brandChannel}
\alias{rbu_md_brandChannel}
\title{基础资料品牌渠道表
t_mrpt2_md_brandChannel}
\usage{
rbu_md_brandChannel(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
基础资料品牌渠道表
t_mrpt2_md_brandChannel
}
