% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_costCenter_directFee}
\alias{rbu_md_costCenter_directFee}
\title{Title}
\usage{
rbu_md_costCenter_directFee(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
\examples{
rbu_md_costCenter_directFee()
}
