% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_costCenter_salesFee_total}
\alias{rbu_md_costCenter_salesFee_total}
\title{Title}
\usage{
rbu_md_costCenter_salesFee_total(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
\examples{
rbu_md_costCenter_salesFee_total()
}
