% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_costCenter_shared_salesFee}
\alias{rbu_md_costCenter_shared_salesFee}
\title{共享成本中心的渠道费用分配比例}
\usage{
rbu_md_costCenter_shared_salesFee()
}
\value{

}
\description{
共享成本中心的渠道费用分配比例
}
\examples{
rbu_md_costCenter_shared_salesFee()
}
