% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_costCenter_total}
\alias{rbu_md_costCenter_total}
\title{成本中心
mrpt2_vw_md_costCenter_total}
\usage{
rbu_md_costCenter_total()
}
\value{

}
\description{
成本中心
mrpt2_vw_md_costCenter_total
}
\examples{
rbu_md_costCenter_total()
}
