% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterData.R
\name{rbu_md_rptItem}
\alias{rbu_md_rptItem}
\title{报表项目基础表
t_mrpt_rptItem}
\usage{
rbu_md_rptItem(variables)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
报表项目基础表
t_mrpt_rptItem
}
\examples{
rbu_md_rptItem()
}
