% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{rbu_normalize_rpa}
\alias{rbu_normalize_rpa}
\title{针对管报的RPA结果进行标准化}
\usage{
rbu_normalize_rpa(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 7,
  FBrand = "自然堂",
  FChannel = "美妆",
  rpa_table = "T_FI_RPA"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{FBrand}{品牌}

\item{FChannel}{渠道}

\item{rpa_table}{rpa所在表,由于进行了版本化可能存在多个表}
}
\value{
返回值
}
\description{
针对管报的RPA结果进行标准化
}
\examples{
library(rbupkg)
# 示例1 使用标准数据
rbu_normalize_rpa() -> data_rpa
View(data_rpa)
# 示例2使用历史版本数据 T_FI_RPA_202107
rbu_normalize_rpa(rpa_table = 'T_FI_RPA_202107') -> data_rpa_202107
View(data_rpa_202107)
}
