% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{rbu_normalize_target}
\alias{rbu_normalize_target}
\title{针对手工管报数据进行标准化处理}
\usage{
rbu_normalize_target(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 7,
  FBrand = "自然堂",
  FChannel = "美妆"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{FBrand}{品牌}

\item{FChannel}{渠道}
}
\value{
返回值
}
\description{
针对手工管报数据进行标准化处理
}
\examples{
library(rbupkg)
rbu_normalize_target() -> mydata
View(mydata)
}
