% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_actual}
\alias{rbu_rollup_actual}
\title{针对历史数据进行处理
mrpt_calc_runAll_actual}
\usage{
rbu_rollup_actual(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 1,
  plan_table = "t_mrpt2_calcPlan_actual",
  src_table = "vw_mrpt_actual",
  res_table = "t_mrpt3_res_actual",
  data_table = "vw_mrpt_actual_count"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{plan_table}{计划表}

\item{src_table}{来源表}

\item{res_table}{结果表}

\item{data_table}{数据表}
}
\value{
返回值
}
\description{
针对历史数据进行处理
mrpt_calc_runAll_actual
}
\examples{
mrpt_calc_runAll_actual()
}
