% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_actual_cumSum}
\alias{rbu_rollup_actual_cumSum}
\title{计算历史数据累计数
mrpt_calc_cumSum_Period}
\usage{
rbu_rollup_actual_cumSum(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2019,
  FPeriod = 2,
  table_data = "t_mrpt3_res_actual",
  table_cumSum = "t_mrpt3_res_actual_cumSum"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}

\item{table_data}{数据表明细}

\item{table_cumSum}{数据表汇总}
}
\value{
返回值
}
\description{
计算历史数据累计数
mrpt_calc_cumSum_Period
}
\examples{
mrpt_calc_cumSum_Period()
}
