% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_budget}
\alias{rbu_rollup_budget}
\title{针对执行预算进行处理
mrpt_calc_runAll_budget}
\usage{
rbu_rollup_budget(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 1,
  plan_table = "t_mrpt2_calcPlan_budget",
  src_table = "vw_mrpt_budget",
  res_table = "t_mrpt3_res_budget",
  data_table = "vw_mrpt_budget_count"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{plan_table}{计划表}

\item{src_table}{来源表}

\item{res_table}{结果表}

\item{data_table}{数据表}
}
\value{
返回值
}
\description{
针对执行预算进行处理
mrpt_calc_runAll_budget
}
\examples{
mrpt_calc_runAll_actual()
}
