% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_res_all}
\alias{rbu_rollup_res_all}
\title{计算结果表,整合当期及累计期间数据
基本不用这个方法了
rds_vw_T_FI_RPA_all}
\usage{
rbu_rollup_res_all(variables)
}
\arguments{
\item{variables}{变量}
}
\value{
返回值
}
\description{
计算结果表,整合当期及累计期间数据
基本不用这个方法了
rds_vw_T_FI_RPA_all
}
\examples{
rbu_rollup_res_all()
}
