% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_res_rpa}
\alias{rbu_rollup_res_rpa}
\title{计划结果计算-RPA数据
从7月开始计算，保留6月数据
mrpt_calc_runAll_rpa}
\usage{
rbu_rollup_res_rpa(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 7,
  plan_table = "t_mrpt2_calcPlan",
  src_table = "vw_fi_rpa",
  res_table = "t_mrpt3_res",
  data_table = "vw_mrpt_rpa_count"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{plan_table}{计划表}

\item{src_table}{来源表}

\item{res_table}{结果表}

\item{data_table}{数据表}
}
\value{
返回值
}
\description{
计划结果计算-RPA数据
从7月开始计算，保留6月数据
mrpt_calc_runAll_rpa
}
\examples{
mrpt_calc_runAll()
}
