% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_res_sumSum}
\alias{rbu_rollup_res_sumSum}
\title{计算累计数_实际数
mrpt_calc_cumSum_Period_res}
\usage{
rbu_rollup_res_sumSum(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2019,
  FPeriod = 2
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
计算累计数_实际数
mrpt_calc_cumSum_Period_res
}
\examples{
mrpt_calc_cumSum_Period_actual()
}
