% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_res_target}
\alias{rbu_rollup_res_target}
\title{计划结果计算-手工管报数据-当期数据
1-6月数据采用target人工计算数据
mrpt_calc_runAll_target}
\usage{
rbu_rollup_res_target(
  conn = tsda::conn_rds("jlrds"),
  FYear = 2021,
  FPeriod = 1,
  plan_table = "t_mrpt2_calcPlan",
  src_table = "t_mrpt_target",
  res_table = "t_mrpt3_res",
  data_table = "vw_mrpt_target_count"
)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{plan_table}{计划表}

\item{src_table}{来源表}

\item{res_table}{结果表}

\item{data_table}{数据表}
}
\value{
返回值
}
\description{
计划结果计算-手工管报数据-当期数据
1-6月数据采用target人工计算数据
mrpt_calc_runAll_target
}
\examples{
rbu_rollup_target()
}
