% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptRolup.R
\name{rbu_rollup_ytdCumsum}
\alias{rbu_rollup_ytdCumsum}
\title{计算累计数，意义不大，整体做的计算
rds_t_mrpt_res_cumsum_rpa}
\usage{
rbu_rollup_ytdCumsum(conn = tsda::conn_rds("jlrds"), FYear = 2021, FPeriod = 6)
}
\arguments{
\item{conn}{连接}

\item{FYear}{年份}

\item{FPeriod}{期间}
}
\value{
返回值
}
\description{
计算累计数，意义不大，整体做的计算
rds_t_mrpt_res_cumsum_rpa
}
\examples{
rbu_rollup_ytdCumsum()
}
